Added segmented Controller
It’s currently working and is reloading the tableView data. But it’s
loading wrong because it overrides the post variable’s array of Posts
so that you can’t filter something several times. To fix I’ll have to
get the data from the db each time or store the whole array of posts in
one large variable